Extract Custom Label to independent class
