+++ MyUi.py
@@ -1,0 +1,11 @@
+from PyQt5 import QtWidgets, QtCore
+
+
+class MyLabel(QtWidgets.QLabel):
+    clicked = QtCore.pyqtSignal()
+
+    def __init__(self, parent=None):
+        QtWidgets.QLabel.__init__(self, parent)
+
+    def mousePressEvent(self, ev):
+        self.clicked.emit()
